feat: Move dialogs into dialogs directory 🧼
