+++ frontend/src/components/dialogs/remove_broker.svelte
