+++ frontend/src/components/dialogs/remove_pipeline.svelte
